requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
ccxt==3.0.75
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.3
pycares==4.3.0
pycparser==2.21
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
urllib3==1.26.15
yarl==1.9.1
